v0.5.0: DAL, FAL and Other Improvements
Note: From now on, the kernel of GarnOS will be reffered to as the Garn Kernel.

Device Abstraction Layer (DAL): Subsystem for Device Abstraction. Implements standard sturctures and functions for Device and Driver Manipulation. Also handles Drives and Partitions and communicates with the FAL for managing filesystems.

Filesystem Abstraction Layer (FAL): Subsystem for Filesystem Management. Garn's shim layer; Implements a standard interface for filesystems.

IDE: Added a driver for IDE Controllers.

VFS: the VFS was replaced by the FAL.

FAT: FAT12/16/32 Identification. Doesn't have any actual functionality for performing file operations. For now it is part of the FAL, but it will be moved to a separate module once I fix a bug inside the DAL.

Some fixes and improvements were also made.